데이터타입 수정 int -> bigint
- 학생, 학습지, 문항, 학습지_문항, 답안 테이블 (총 5개) 에서 id들을 int에서 bigint로 수정
- java에서 int를 Long으로 수정

+++ diagnosis/sql/create.sql
@@ -35,7 +35,7 @@ CREATE TABLE USERS (
 
 -- 학생 테이블
 CREATE TABLE STUDENTS (
-	student_id	INT auto_increment,
+	student_id BIGINT auto_increment,
 	student_name VARCHAR(20),
 	student_phone VARCHAR(20),
 	student_birthdate DATE,
@@ -48,18 +48,18 @@ CREATE TABLE STUDENTS (
 
 -- 학습지 테이블
 CREATE TABLE TESTS (
-	test_id	INT auto_increment,
+	test_id	BIGINT auto_increment,
 	test_name VARCHAR(20),
 	test_comments VARCHAR(200),
 	test_timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-	diagnostic_test_id INT,
+	diagnostic_test_id BIGINT,
 	PRIMARY KEY (test_id),
 	FOREIGN KEY (diagnostic_test_id) REFERENCES TESTS (test_id)
 );
 
 -- 문항 테이블
 CREATE TABLE ITEMS (
-	item_id	INT auto_increment,
+	item_id	BIGINT auto_increment,
 	item_answer	VARCHAR(20),
 	item_image_path	VARCHAR(255),
 	concept_id INT,
@@ -69,9 +69,9 @@ CREATE TABLE ITEMS (
 
 -- 학습지-문항 테이블
 CREATE TABLE TESTS_ITEMS (
-	test_item_id INT auto_increment,
-	test_id	INT,
-	item_id	INT,
+	test_item_id BIGINT auto_increment,
+	test_id	BIGINT,
+	item_id	BIGINT,
 	test_item_number INT,
 	PRIMARY KEY (test_item_id),
 	FOREIGN KEY (test_id) REFERENCES TESTS (test_id),
@@ -80,9 +80,9 @@ CREATE TABLE TESTS_ITEMS (
 
 -- 답안 테이블
 CREATE TABLE ANSWERS (
-	answer_id INT auto_increment,
-	student_id	INT,
-	test_item_id INT,
+	answer_id BIGINT auto_increment,
+	student_id BIGINT,
+	test_item_id BIGINT,
 	answer_code	INT,
 	answer_probability DECIMAL(5,2),
 	answer_timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
